+++ Changelog.md
@@ -3,3 +3,9 @@ CHANGELOG
 
 0.1
 - added functionality for it to retrieve data from bitstamp.com
+
+0.2
+- Retrieves bitstamp info from ticker
+- Displays most recent bitcoin information
+- writes ticker info to a .csv file
+- added button that allows opening of .csv file from windows.